Handle gulp not being available
Log an error if gulp is not available. Checking for gulp tasks runs
gulp. If gulp is not available then a Win32Exception is thrown
indicating the gulp application is not available. This exception is
logged using the TaskRunnerLogger so it is available in the Task
Runner Explorer Output tab. Also return a TaskRunnerErrorNode so
the Show Task Runner Explorer Output menu item is displayed for the
tree nodes in the Task Runner Explorer window.